+++ CHANGELOG.md
@@ -102,3 +102,8 @@ Initial release of `slim-tools` with the following features:
 - - `context-restrictions.ts` checks for functions being used in the wrong context, e.g. the wrong callback or wrong WF/nonWF mode
 - - `initialization-rules.ts` ensures the simulation is properly initialized
 - - `interaction-queries.ts` ensures interactions are only used after they're evaluated
+
+## [0.0.14] (CAT)
+- Added the signature help provider (`signature-help.ts`) to provide hints on parameters in functions/methods
+- Added the code actions provider (`code-actions.ts`) to provide quick fixes for common errors and simple refactorings
+- Improved the validation service to avoid throwing errors while a line is being typed